pyproject.toml
[tool.poetry]
name = "hmeg"
version = "0.1.6"
description = "Generator of exercises for practicing speaking for language learning."
authors = ["Yury Choi <yurytsoy@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/yurytsoy/hmeg"

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
fire = "^0.6.0"
inflect = "^7.0.0"
jsonschema = "^4.20.0"
kenlm = "^0.3.0"
language-tool-python = "^2.9.4"
nltk = "^3.8.1"
numpy = "^2.0.0"
openai = "^2.15.0"
pandas = "^2.2.2"
python-levenshtein = "^0.27.1"
pyyaml = "^6.0.0"
sentencepiece = "^0.2.1"
spacy = "^3.8.7"
toml = "^0.10.2"
torch = "^2.8.0"
torchvision = "^0.23.0"
transformers = "^4.56.1"
python-dotenv = "^1.2.0"
orjson = "^3.11.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
